/battle_list 구현 완료 및 오류 수정 (수정내용 : DAO에서 받는 부분의 첫번째 인자를 int가 아닌 long이어야 함, 유닛 테스트에서 setting.realrank 폴더 찾는 경로 수정)

+++ src/realrank/battle/BattleManager.java
@@ -31,7 +31,7 @@ class BattleManager {
 
 		ArrayList<BattleInfo> battleList = new ArrayList<BattleInfo>();
 		queryResults.forEach(result -> {
-			battleList.add(new BattleInfo((Integer)result.get(0), (String)result.get(1),
+			battleList.add(new BattleInfo((long)result.get(0), (String)result.get(1),
 					(String)result.get(2), (Date)result.get(3), (Date)result.get(4),
 					(Integer)result.get(5), (String)result.get(6)));
 		});
@@ -50,7 +50,7 @@ class BattleManager {
 
 		ArrayList<BattleInfo> battleList = new ArrayList<BattleInfo>();
 		queryResults.forEach(result -> {
-			battleList.add(new BattleInfo((Integer)result.get(0), (String)result.get(1),
+			battleList.add(new BattleInfo((long)result.get(0), (String)result.get(1),
 					(String)result.get(2), (Date)result.get(3), (Date)result.get(4),
 					(Integer)result.get(5), (String)result.get(6)));
 		});
@@ -69,7 +69,7 @@ class BattleManager {
 
 		ArrayList<BattleInfo> battleList = new ArrayList<BattleInfo>();
 		queryResults.forEach(result -> {
-			battleList.add(new BattleInfo((Integer)result.get(0), (String)result.get(1),
+			battleList.add(new BattleInfo((long)result.get(0), (String)result.get(1),
 					(String)result.get(2), (Date)result.get(3), (Date)result.get(4),
 					(Integer)result.get(5), (String)result.get(6)));
 		});
